--- src/main/kotlin/Analista.kt
@@ -1,12 +1,0 @@
-class Analista(
-    nome: String,
-    cpf: String,
-    salario: Double
-) : Funcionario(
-    nome = nome,
-    cpf = cpf,
-    salario = salario
-) {
-    override val tipo = "Analista"
-    override val bonificacao = salario * 0.1
-}
